Allow blocks (procs/function) to share names with non-blocks. Previously all block/function/proc names had to be unique among the names of all block/function/proc names. With the change a block may share a name with a proc, for example, but not with other blocks. The motivation for this change is that we'd like the name of blocks to bear the name of the emitted verilog module and often that is the same as the source function name.
An alternative considered might be tagging blocks with the desired name of the emitted Verilog module as a field separate from the name of the block itself. However, the addition complexity of adding a field isn't justified given that different ways blocks/procs/functions are used and invoked avoids ambiguity between such constructs sharing the same name.

PiperOrigin-RevId: 380064177