[DSLX] Fix parametric instantiation of enum values.
This led to us querying in the code what the bit width of the type was for a
parametric binding during parametric instantiation. (That is, on arbitrary
ConcreteTypes instead of just `bits`-style BuiltinTypes.) To make that happen
cleanly needed the world outside of the ParametricInstantiator to deduce the
types and pass them in along with the ParametricBinding AST nodes, so broke out
a type that decorated the ParametricBinding with the deduced ConcreteType and
we pass that into parametric instantiation routines now.

PiperOrigin-RevId: 381057765