Add SingleValue channels. SingleValue channels are a new type of channel. A SingleValue channel only holds a single value at a time, and receives read this value non-destructively. Sends overwrite the current value. This differs from Streaming channels which have FIFO semantics. Existing Port and Register channels are obsolete and will be removed in follow-up changes.
Currently not used, but uses will be added in later changes.

An example use case for single-value channels might be a two-way mux (two data inputs, one data output, and a selector input). The inputs and outputs might be streaming channels, while the selector for the mux might be a single-value channel. The selector is written once, while many values might stream through the mux through the data channels.

PiperOrigin-RevId: 381379738