[DSLX] (Pre)factor out a StructType from ConcreteType.
Since structs were added after tuples they were bolted in to the tuple type
(and in many ways they're polymorphic with tuples, but the is-a and variant
guts is too strong a coupling, more than was necessary).

This also uncovered an issue where, because we were treating structs and tuples
similarly, you could use tuple indexing on struct values, made an
`Vdslx/tests/errors` test to show that's considered an error.

PiperOrigin-RevId: 381511658