[DSLX] Flag an error when a syntactically-duplicate match pattern is detected.
Only the first would ever match, since the environment cannot change from
pattern match arm to pattern match arm.

Note that expressions are not allowed in match-pattern position, just patterns,
which today are roughly "names or constants / constant-references, and tuples
thereof". This change adds text to the DSL reference because users might think
we do an _equivalence check_ of all constant values in the pattern, which we
don't, it's _only_ syntactic equivalence as an opportunistic check (since
there's not an easy way today to check whether patterns are equivalent when the
constexprs are resolved). Filed google/xls#459 to track the potential for more
general pattern equivalence.

PiperOrigin-RevId: 385207239