[DSLX] Fix mapping of parametric instantiation in local module.
Before we were clobbering the `CurrentTypeInfo()` in the interpreter with the
mapped_fn's module's type information. But if it's a parametric instantiation,
we actually can't use the module's type info, we need to use the instantiated
type info.

Notably putting a `x << 0` in a function forces the DSL interpreter to resolve
the `0` in the type info to figure out what width to use, kind of handy for
finding type information issues.

* Switches various TypeInfo functions to take references to reflect
  non-nullability.
* Fix for newly added sample.
* Added missing assertion of main() in existing map.x sample.

PiperOrigin-RevId: 386056355