Moving the ubiquitous "additional_search_paths" argument inside ImportData.
ImportData is a natural location for this data, as having *different* search paths per-module could lead to odd resolution problems (which foo.x should we import?), especially as we're interested in building one top-level module, which reasonably should have a single view of its dependency space.

Cleans up/simplifies a LOT of APIs.

PiperOrigin-RevId: 386320002