Add data structure which gathers registers and register nodes together. Block::GetRegisterNodes returns a map from register name to a struct containing the respective RegisterWrite and RegisterRead nodes.
Also, fix issue where port names could collide with interior node names and result in port names with numeric suffixesi which is undesirable because port names are part of the observable interface in the emitted Verilog module. This fix renames interior nodes out of the way of any newly added port names.

PiperOrigin-RevId: 391551697