Associate RegisterRead and RegisterWrite ops with registers in block. Keep a map from Register* to the associated RegisterRead/Write operations. Keep data stucture consistent in AddNode/RemoveNode. This replaces a previous mechanism which returned a map generated on demand.
The mechanics of associating the Register/RegisterRead/RegisterWrite constructs together is a bit clunky, but with the checks in place and with the verifier changes the risk of any incoherence should be minimal. A (radical) alternative would be to have a register xls::Node which encapsulates the write, read, and the register itself into a single construct. Unfortunately, this would require a cyclic IR which would be a very significant change with unknown implications.

PiperOrigin-RevId: 393429211