Updating AST definition and parsing for new-style procs.
This change is the first in a series to separate out config and iteration components of procs into separate "code" units. As part of this change, the "iota" test case is been updated to use the new syntax but was removed from the examples BUILD file, as the rest of the DSLX stack updates still need to land.

In detail, this change "desugars" a proc declaration into two functions: a [constexpr] "config" function which defines proc constant members (including channels), creates channels, and can spawn other procs, and a "next" function which determines per-tick behavior. Subsequent changes will add typechecking and lowering to IR.

The composition of this and the next few changes may be odd, as this feature was developed monolithically, but it needs to be separated to avoid presenting a colossal patch for review. As always for a change of this magnitude, comments and questions and suggestions are welcome.

PiperOrigin-RevId: 399806801