Updating the inlining pass such that coverpoints and asserts aren't duplicated when functions containing such are inlined < 1x.
To do this, we add the ID of the invoking node to the label attached to the cover or assert. This will result in a single coverpoint being "spread out" over several coverpoint instances when lowered to IR - this may require postprocessing to re-aggregate the data, depending on user needs, but is necessary, as covers/asserts must have unique names in Verilog.

PiperOrigin-RevId: 401337174