Adds lowering to IR for new-style procs.
This completes support for the first end-to-end example, [re-]added to the example build rules here. Subsequent changes are planned to A) verify constant proc member support and B) update Proc config function member setting syntax to not use Let binding but to instead use the outputs of that function to set members (in the order in which they were declared).

Still, this is a milestone worth checking in.

PiperOrigin-RevId: 402882798