Distributed routing algorithm for multiple paths between endpoints.
The algorithm selects the paths with minimum hop between two endpoints. If there are multiple paths with equal min hops between two endpoints, then the input port to the output port assignment are performed using a modulo technique. For example, a router with six input ports and three output ports where the output ports have min hops for a given endpoint pair, the asignment is as follows:
 Output Port Index   | Input Port Index
 ----------------------------------------
           0         |     0, 3
           1         |     1, 4
           2         |     2, 5

PiperOrigin-RevId: 405988139