netlist_interpreter_main: handle input ports in declaration order
YoSys, when emitting netlists, often has input-port declarations in a different
order than the one in the module header, eg:

	module ifte(i, t, e, out);
		input [7:0] e;
		input i;
		output [7:0] out;
		input [7:0] t;

When providing input, the --input string should follow the order parameters are
listed in the module declaration, rather than the order of write declarations.
This patch does a few things:

a) Adds information about the declaration order of aggregated inputs, their
widths, and their directions, while parsing a netlist.  In the example above,
we save the information that input `i' is followed by input `t', then input
`e', and output `out', with their respective widths being 1, 8, 8, and 8 bits,
and their directions being respectively in, in, in, and out.

b) Uses that information in netlist_interpreter_main to compute the correct
mapping from the --input values to the input ports of the netlist.

c) Removes the string "Results: " when reporting results.

d) Adds a simple test target for the netlist interpreter.

PiperOrigin-RevId: 407250682