Add support for assign statements.
Verilog netlists generated by YoSys and ABC from input files that have gone
through the XLS toolchain may contain simple assign statements.  The two cases
that seem to generate such assign statements are:

1. C++ types that are wider than what is needed.  For example, computing the
square root of an 8-bit number will give you a 4-bit number, but C++ does not
have a convenient way to express a nibble.  If the return type is also declared
to be 8-bit, then the upper nibble will never be used.  The toolchain
accommodates this by assigning zeros to the additional wires.

2. C++ nonconst references in XLScc map to an assignment of the input to the
output wires.

This patch adds support for a subset of the assign statement to handle the
above two cases.

PiperOrigin-RevId: 407253285