Updating TestProc syntax and parsing for recent design changes.
I've been speaking with Mark about proc testing design, and since IR channels can only be used by a single send and receive node, it makes sense to have the same semantic in the DSL. That means, though, that TestProcs need to have their `next()` functions run multiple times (to potentially inject multiple stimuli into the network), which means that A) they need to hold state, and B) they need a means to tell the interpreter that they've finished executing.

This changes enable these by A) enabling the specification of initial values to a TestProc and B) requiring that a TestProc's `config()` function accept a boolean channel to indicate test completion (as well as pass/fail status).

PiperOrigin-RevId: 408635954