[public] Add a test that reflects public API consumers.
This test directory is not considered part of internal XLS and thus has the
same access perms as a "public" user in the repo, so it reflects the headers
and access capabilities available via the public APIs and tests them for
functional usability.

Since we'd expect most functionality to be covered via internal tests, this
will likely just hold some key smoke tests that functionality is accessible and
demos of how it can be used from a "public" user location.

PiperOrigin-RevId: 408932946