Adding functionality towards "push-button" LEC execution.
This adds an "--auto_stage" flag, which uses heuristic(s) to determine if the current design should use full or staged LEC. Such a heuristic might be the presence of a mul/div operation in the design or its overall size. This flag eliminates the need for the user to make such a decision - instead, the user provides the tool with all relevant info (i.e., the schedule along with the normal data) and it handles the rest.

PiperOrigin-RevId: 408996144