Avoid flattening associative ops if it increases op count. Expressions of n-ary associative ops are flattened in arithmetic simplification. Example:
  Op(x, Op(y, z)) => Op(x, y z)

Currently this is done unconditionally whether or not the nested op (e.g., `Op(y, z)`) has other uses. This results in additional operations being added to the graph and the optimization advantage in these cases is not clear. This change does not flattened nested operations which have other uses.

PiperOrigin-RevId: 410853732