Generalize table switch pass. Generalize table switch pass in two ways:
(1) Support neq comparisons in the select chain. Previously only eq was used. However, `sel(a==b, [x, y])` is equivalent to `sel(a!=b, [y, x])` so it makes sense to take advantage of this property.

(2) Allow more than one "hole" in the table which is filled by the fall-through literal value. Previously, the transformation kicked in only if the fall-through value at the end of the chain corresponded to exactly one value of the indexing value. However, that is needlessly constraining. Now as long and the fall-through value fills less than half of the table the optimization kicks in.

PiperOrigin-RevId: 411700611