[build] Helper macro for checksumming files.
As projects cut releases or freeze, it's helpful to know that the target is
never changing without having to actually check in generated Verilog code.
Users might use this macro to help enable confidence that there is neither:

a) non-determinism in the toolchain, nor
b) an accidental dependence on a non-released toolchain (e.g. an accidental
   dependence on top-of-tree, where the toolchain is constantly changing)

PiperOrigin-RevId: 413701250