XLS: Templatize the Interpreter, Parser, Netlist, NetRef, Module, Cell, CellLibrary.
This is a large atomic change that converts most classes under
//xls/netlist/... to generic ones, templatized on a type called EvalT, that acts
like a boolean, with the following properties:
a) It may or may not actually evaluate to a bool.
b) It may or may not be represented internally as a boolean.
c) It does implement all operations on bool values.

The classes Parser, Netlist, Module, NetDef, Cell, CellLibrary, and their
related supporting classes have been renamed with the prefix "Abstract".  The
original names are type-aliased to the default value of bool.  Code using them
does not change.

The objective of this change is to prepare the netlist interpreter and its
related infrastructure to evaluate Verilog netlists over such more-generic bool
stand-in types.

Major components:
1) Templatize classes on a generic value Eval that acts like a boolean, but may or
   may not evaluate to bool; rename with prefix "Abstract".
2) Construct 0-value and 1-value using Eval-template constructor.
3) Pass Eval-0 and Eval-1 as ground truth values.
4) Also add an alias for the map from NetRef to Eval.
5) Convert code to use templatized versions.

Also:
6) Pass processed_wires by reference instead of by pointer to InterpretCell for
   consistency with the rest of the code.
7) Make more components visible to XLS users.

Tests:
8) Use a simple cell library (and, or, not);
9) Use a netlist that computes XOR;
10) Test with both function and statetable implementations;
11) Test with bool and an opaque value representing bool.
PiperOrigin-RevId: 415627088