XLS: Speed up the netlist parser
Netlists consist mostly of wires and cell instantiations; parsing them involves
a lot of repetitive lookups by name.  Replace the cell and wire by-name lookups
with hash-map ones.  Also, precompile the RE2 matcher for Verilog numbers so
that we don't do the heavy lifting in the hot path for each number.

The speed up for parsing a large netlist:
-- Before: 266 seconds
-- Now: 0.20 seconds (1330x improvement)

PiperOrigin-RevId: 415661165