XLS: Define Cell::OutputPin with an eval function.
Provide the ability to pass evaluation functions for each output pin of each
cell in the netlist.  When an evaluation function is provided for a given
cell's output pin, it is invoked to compute the value of the cell for that
output pin, instead of parsing and interpreting the actual cell's function
definition or its state function.

Test the evaluation functions in the netlist interpreter_test.

PiperOrigin-RevId: 416130885