XLS: Specify cells that use a wire as an input
In addition to keeping track of all connected cells, for which an
AbstractNetDef may serve both as an input and an output, add a list of cells
for which the given wire is an input.  This is the most common case, and
explicitly tracking it allows us to avoid checking if a wire is an output when
interpreting a netlist, which provides a sizeable speed improvement (at the
cost of lugging more state in AbstractNetDef).

PiperOrigin-RevId: 420786223