XLS: Multi-thread the netlist interpreter
Add an optional parameter to AbstractInterpreter's constructor for the number
of threads in the thread pool.  Default is zero, disabling multithreading.

In InterpretModule, when a cell becomes ready to be interpreted, we check the
state of the threadpool.  If there are threads ready for more work, we dispatch
the cell and its inputs to by queueing it up and waking a worker thread.

The need to dispatch work only if threads are ready has to do with the
recursive nature of InterpretCell, which may invoke InterpretModule.  When this
happens, the call to InterpretModule will take place in its own dedicated
thread, farming InterpretCell work out to other threads as needed.  If we do
not check the thread pool status, we may get into a deadlock where all worker
threads are handling InterpretModule calls and none is left to handle
InterpretCell.  This check also acts as crude load-balancing, making sure
InterpretModule also does some work when all threads are busy.

A more intelligent check would be to handle work locally only if all pool
threads but one are currently busy handling InterpretModule and the next cell
is a module.

PiperOrigin-RevId: 421053506