[BUILD] Fix breakage caused by new Bazel version and depending on an old gRPC.
I believe something about the way the gRPC rules were set up in the old
revision of grpc was causing the error (surfaced by Bazelisk switching us
to Bazel 5).

Upgrading gRPC it has a reference to a newer Cython so we can drop our patch,
but we do need to configure Python ahead of loading gRPC in the WORKSPACE flow.

PiperOrigin-RevId: 423227915