XLS: Add optional is_declared_as_char to XLSCC metadata
C++ chars are neither signed nor unsigned.  Add an optional metadata field and
set it to true when a char declaration is encountered. Set it to false for
explicitly-signed- or explicitly-unsigned-char decalarations.  Do not set it
altogether for integer-type declarations of bit widths other than 8.

This is useful in being able to tell char from u/int8_t declarations when only
looking at metadata.

PiperOrigin-RevId: 426941850