[DSLX] Helper macro for DSLX language-level tests.
Convenience macro for DSLX language-level tests and associated targets.

Also includes minor improvements to some of the error messages encountered
while enabling more test targets.

Often when we define a DSLX language-level test we also want to know that file
converts to IR and evaluates as IR appropriately.

Note: these are not user-facing, these are just for convenience of internal
development in specifying the various targets we usually want stood up as part
of defining a DSLX language-level test.
PiperOrigin-RevId: 427021623