XLS: Ignore wire declarations for already-existing port declarations.
Certain tools, such as YoSys as of commit
56e7791760ce67cb1831691460b50bf73a4f5117 "verilog backend: Emit a `wire` for
ports as well. by Marcelina Kościelnicka <mwk@0x04.net>", will emit a wire
declaration for an already-declared input/output port.  When a duplicate net
declaration is encountered, error out only when the net is kWire.  For kInput
and kOutput, ignore the duplicate.

PiperOrigin-RevId: 427227937