[DSLX] Start of bytecode interpreter support for Match construct.
* Adds pop bytecode, fixes dup bytecode.
* Start of laddered matching -- only works for single-pattern-in-a-given-arm at
  the moment with a trailing wildcard pattern match.

## (Side note on how we'll handle pattern-created bindings)

I'm not sure that we have a construct in the bytecode emitter yet that can
scoped-ly bind a name, so that'll probably be the challenge for the next
change; e.g.

```
let x = (u32:42, u32:64);  // here x is a tuple
let e = match x {
  (x, y) => x,  // here x is a u32
};
let _ = f(x);  // here x is a tuple again
```

We _could_ use rspringer@'s approach for `map` making a `BytecodeFunction` for
that particular scope and pass things in as args, or we can traverse the AST
numbering the NameDefs with their slots (and then `NameRefs` point at their
`NameDefs` so you always know what slot to look up). Seems like that'd be
faster, likely just requires an AST pre-pass, and is pretty common in bytecode
schemes, but warrants some discussion.

PiperOrigin-RevId: 428057274