Add proc inlining pass. The proc inlining pass inlines all procs into the top-level proc. Each proc is managed as a separate thread of execution in the top-level proc managed by an activation bit which flows through the proc side-effecting operations.
Proc inlining is a valid transformation for networks of procs for which the receive operation of the channel is always ready when the respective send executes. That is, proc networks which don't require fifos.

Currently only works for streaming channels. Single-value channels will be added in a followup CL. Also, not yet connected to the pass pipeline. This will be done in a followup CL which adds it behind a pass option.

PiperOrigin-RevId: 428152746