Several improvements to proc inlining. (1) Connect proc inlining to the standard pass pipeline. It is enabled by fields on PassOptions because we don't want to run it unconditionally. This is not a great way to enable it, but is workable for now. It runs after useless IO removal pass because conditional send/receive on single-value channels is not supported and some blocks require the dead io removal pass to run first in order to translate conditional send/receive with always-true predicates into unconditional send/receives. This may not be the best place for this pass optimization-wise.
(2) Support inlining of procs with external IO. Previously, only the top proc could do external IO. In fact, that was the way the top proc was identified. Now the top proc is specified via a field on PassOptions. When Package::EntryFunction is generalized to support procs (and renamed to "top") we can eliminate this field.

(3) Little cleanup of proc inlining which makes the activation chain an explicit data structure.

(4) Bug fix related to send/receive cloning.

PiperOrigin-RevId: 428519587