XLS: Support more-flexible assign statements in the netlist parser and interpreter.
Support for assign statements does not quite cover the range of syntax seen in
netlists emitted by YoSys.  This change allows for that more general syntax.
When encountering an assign declaration, we break the LHS and RHS into
individual wires, from MSB to LSB.  Then, starting with the LSB, we start
assigning the RHS to the LHS.  This permits the loose mapping of any
combinations of wires and number-literal values.

Given a list of wires representing the LHS and RHS of the assignment, each
complex wire is broken into its individual bits; single wires are represented
by their name only. (Number literals and input ports are not allowed to be
present on the LHS.)

For example, consider in the following port (equivalently, wire) declarations:

  input a;
  input [2:0] b;
  output [3:0] c;
  output d;

And the following assignments:

  assign c = { a, b }
  assign d = a;
  assign { c, d } = { a, b[2:1], 1'b10 };

The LHS for the assigns would each become a vector of strings with values as
follows:

  "c[3]", "c[2]", "c[1]", and "c[0]".
  "d"
  "c[3]", "c[2]", "c[1]", "c[0]", and "d".

For the RHS:

  "a", "b[2]", "b[1]", "b[0]"
  "a"
  "a", "b[2]", "b[1]", "1", "0"

These would be matched up wire by wire from right to left.  Note that the LHS
and RHS need not be of the same size.  If there are more terms in the LHS, then
the extraneous most-significant ones will be unassigned.  If there are more
terms on the RHS, then the most-significant values would not be assigned to
anything.

This change also allows for assign statements where wires can be present in
both the LHS and RHS (previous implementation required only literals on input
wires to be on the RHS, and only output wires to the on the LHS.)

PiperOrigin-RevId: 428861350