Adding `trace!` and `trace_fmt!` support to the bytecode interpreter.
This change also updates `trace_fmt!` to always add a trailing newline to its output.

Note that one will occasionally see duplicate tracing output. This is temporary, and due to the fact that tracing calls may be interpreted during typechecking as well as during runtime (for certain functions). This will be fixed during the transition to the new interpreter.

PiperOrigin-RevId: 431762635