Removing the requirement for the last match arm to be a wildcard from the bytecode emitter.
In addition, this adds support for match arms with different numbers of elements than the input values. Match statements like the following are now supported:

```
  let x = (a, (b, c));
  match x {
    (u32:1, (u32:2, u32:3)) => u32:0,
    (u32:2, _) => u32:1
  }
```

This is the last non-proc functionality missing from the bytecode interpreter (that I know of), so migration off of the AST-walking interpreter can now begin.

PiperOrigin-RevId: 432479088