Modifies the typechecker to not "backtrack" when encountering a parametric function.
Previously, the typechecker would report an error and re-start typechecking when a parametric function was encoutnered in a call stack. This system was a bit hard to grok and proved difficult to modify (changes are needed to fully support Procs).

This change alters the behavior such that parametric calls are typechecked immediately as they're seen, with no backtracking.

The performance of `bazel test -c opt xls/...` is improved roughly 17% by this change.

PiperOrigin-RevId: 439451516