Make proc inlining work with arbitary proc network topologies. Fan-in and fan-out of tokens are now supported and a bunch of other latent issues are corrected. All proc networks that are executable with single-element FIFOs for the streaming channels should be inlinable now. There is also an issue where if a receive->send has a data-dependency but not token dependency then the result will be incorrect. A disabled test is added for this case.
This change saves data received on *all* channels both internal and external to proc state (which later become flops). This results in a significant increase in the number of flops generated after proc inlining. Later CLs will add proc state optimizations which should remove unnecessary flops. This separation of concerns (naive proc inlining and later proc state optimization) is necessary to keep proc inlining complexity manageable.

To avoid the QoR hit on blocks which work with the existing proc inlining pass until the proc state optimization is done, this new functionality is enabled via a temporary flag. The old proc inlining pass delegates to the new one when this flag is set.

PiperOrigin-RevId: 440135233