Adds TestProc execution support to the bytecode interpreter.
The basic idea is that a special BytecodeInterpreter (ProcConfigBytecodeInterpreter) traverses the network and instantiates a `Process` for each encountered proc, which are then collected and returned to the top level. After that, the top-level function `RunTestProc` (in run_routines.cc) ticks the network until the special terminator channel has data.

PiperOrigin-RevId: 444659894