Removing the automatic DSLX test generator.
Unfortunately, the test generator is built very tightly against the old AST-walking interpreter and isn't easily adapted to the new bytecode interpreter. Since the old interpreter is on its deathbed, this is problematic. In addition, it might make more sense to target the IR, since that'd enable us to target xlscc-generated designs as well. Taken together, it's probably time and ok to delete the existing test generator and to rewrite one against the IR (which will be less of a moving target, anyway).

PiperOrigin-RevId: 444665682