Fix a couple issues in block conversion related to multiple state elements. Two cases were mishandled in block conversion: (1) a node which is the next state for multiple state elements, and (2) a next state node which is not dependent upon the corresponding state element.
Also change the name of the uint64_t overload for sequential block interpretation so it is no longer an overload. This overload was differentiated by the type of the non-streaming input argument. But if there were no non-streaming input elements then ideally {} could be passed as that argument, but that resulted in an ambiguous function call.

PiperOrigin-RevId: 445207018