Change select generation for assignments. This will allow flexible loop unrolling without an explosion of complexity.
Previously, the condition was accumulated as scopes were entered,
and a select was generated immediately upon a conditional assignment.

For example:

int x = 10;
if(a) { if(b) { x = 20; } }

Generated:

x = (a&b) ? 20 : 10;

Now, however, selects are generated upon exiting a scope (popping the
TranslationContext), and using only the condition added to that specific scope,
or the "relative condition".

Thus, two selects would be generated in the sample.

First, the assignment itself is direct:
x = 20

Then, on exiting the if(b) scope, a select is generated:
x = b?20:10

Finally, on exiting the if(a) scope, a second select is generated:
x = a?(b?20:10):10

This is more complex for this sort of trivial case, but the direct assignment in the inner scopes greatly simplifies more complex cases. For example:

int x = 10;
int y = 20;
if(a) {
 x = 100;
 y = x;
}

Previously, the x = 100 would generate a select:

x = a?100:10;

Then the y = 200 would generate another select, incorporating that one:

y = a?(a?100:10):20;

Combined with loop unrolling, this quickly snowballs.

Now, because the selects are generated late, the result is like this:

x = a?100:10;
y = a?100:20;
PiperOrigin-RevId: 445438670