Adds ahead-of-time (AOT) compilation support to XLS.
This system, as currently implemented, causes the JIT to produce object code, which is written to disk. A header file and source file are generated as well - the generated source needs to marshal the input [XLS] capital-V Values into the packed bit buffers used by the JIT'ed code.

This has been tested by manual creation of a program that invokes fp32_add_2. Subsquent changes will integrate this into the build system and add in-system tests.

Part of issue #344.

PiperOrigin-RevId: 447835839