[JIT] Support formatting data operands in trace IR op.
Fixes google/xls#403 -- we still want to remove the older `trace!` builtin in
favor of `trace_fmt!` as well as expand the formatting mini-language support
translating to Verilog but those will be tracked in a separate issue, basic
support is present in JIT execution mode now.

Reuses deserialization-from-native-layout code in the JIT runtime, stores to
stack storage via alloca and passes that to the formatting-step-handler
JIT-runtime-helper function.

PiperOrigin-RevId: 448008113