DSLX interpreter: Relaxing constraint that NameDefs must be seen by incrementing slot index.
Previously, it was assumed that slots would be seen during interpretation strictly as observed during bytecode emission, such that we would only ever have to create a single slot at a time in the interpreter. Unfortunately, that's not the case - NameDefs created inside conditional expressions may never be observed, meaning that subsequent slot accesses would trigger a failure: if we only create a single slot at a time, then an access to slot N would fail, as slot N-1 would not have been created.

PiperOrigin-RevId: 449009910