Fixes missing stack pop when handling out-of-bounds width slice.
It's valid for a width slice to specify a start position outside the size of the value to be sliced: in that case, a zero value should be returned. Prior to this change, the value to be sliced was never removed from the stack - instead, we detected the out-of-bounds case, pushed the zero on the stack, and moved on. This change fixes that and pops the value to be sliced, even in the OOB case.

Fixes #630.

PiperOrigin-RevId: 452550448