Fix crasher 2649 and enable some crashers which no longer fail. This required small DSLX changes to some because the language has evolved slightly.
The underlying problem with 2649 was a corner case where tuple operations are codegen-ed via a separate path which didn't honor the need for the corresponding Verilog value to be a named expression. This resulted in indexing of an inline expression. For example, a three element tuple and a corresponding tuple-index operation was previously emitted as:

assign tmp1 = {a, b, c}[4:1]

This is invalid Verilog. Now this might be emitted as:

assign tmp0 = {a, b, c}
assign tmp1 = tmp0[4:1]

PiperOrigin-RevId: 452555181