DSLX: Adding backrefs to AstNodes.
This will make it possible for constexpr evaluation to be further decoupled from type deduction: currently, when we hit a NameRef, we trace it back to its NameDef. If that NameDef hasn't been evaluated, then we can go no further. With this change, we can find out where that NameDef is defined (even if it lacks a populated `definer`, as will be the case for NameDefTree-embedded NameDefs) and evaluate that tree, and so on.

This required several related changes:
 - ConstantDef nodes were removed from Let nodes; instead, Let nodes now have an `is_const` member that denotes if their RHS is const.
 - Prohibition of destructuring assignment in const declarations is now enforced and reported to the user.
 - Proc `config` and `next` functions now "share" the parent proc's ParametricBindings: only the parent proc owns them as a child, but its constituent functions reference that copy.

Validation was added to verify the appropriate parent/child relations now hold across the AST.

PiperOrigin-RevId: 453222242