Initial steps to decouple constexpr evaluation from typechecking/deduction.
As part of the longer-term plan to introduce unrolled for! macros (a la Rust), it'll be necessary to do some constexpr evaluation up-front (in order to determine the loop count, for example). The first step in doing this is to begin to separate constexpr evaluation from being so tightly coupled to typechecking: prior to this change, constexpr evaluation was triggered upon completion of type deduction of every node.

With this change, most? much? constexpr evaluation is performed on demand, when a constexpr value is needed, but some exceptions remain. Subsequent changes will continue the separation, leading to constexpr evaluation as an independent "pass" over the AST.

PiperOrigin-RevId: 453236401