Add new "priority select" op.
`priority_sel()` is similar to `one_hot_sel()`, but instead of `OR`-ing each selected case, it chooses the first selected case. This adds the operation to the IR.

The existing `one_hot_sel()` naturally maps to an efficient circuit, but its semantics make it an awkward fit for source-level constructs as wells as analysis and optimization. Mixing multiple inputs is rarely desired from the source and complicates analysis. Priority select may be a more useful operation as it more closely mirrors DSLX match statements or C/C++ `if/else if/else` chains. See google/xls#547 for more discussion.

PiperOrigin-RevId: 453780962