Enabling typechecking of top-level procs whose config functions take channel params.
This enables IR conversion of the same.

This change is necessary because converting a construct to IR requires that it has been typechecked, but prior to this change, any proc with config function params could not be typechecked without a driving instantiation.

Naively enabling typechecking of such procs was problematic, because DeduceSpawn requires constexpr eval of all arguments, but a top-level proc has nothing to provide such arguments. Thus, only channels are allowed as arguments to top-level procs, and dummy channels are created in DeduceParam.

With that in place, interpretation would then be broken, because a TypeInfo recurses through its parent to locate items and constexpr values, and so a proc instantiated by a TestProc or other proc would only ever find the dummy channels created by DeduceParam, breaking network instantiation. To fix that, top-level procs are typechecked in a child TypeInfo of the top-level module TypeInfo, instead of that top-level TypeInfo itself.

PiperOrigin-RevId: 458523040