Make more general mechanism to override codegen for various ops.
Previously, there were format string fields for assert and gate ops and the logic to override the default behavior with a format string existed within the operation's Emit*() function. For google/xls#645, I'm adding functionality to use a format string for a new operation and the way the format string is injected into vast is different- unlike asserts and gates, we want to instantiate a user-specified module for `umulp`/`smulp`.

Rather than making another format string field and Emit*() function with special behavior, I thought it would be better to generalize how operations can be overridden. This change adds a map from `Op` -> `OpOverride`. When an Op is being codegen'd, it first checks for an entry in the map. If there is an `OpOverride`, it calls the override's `Emit()` function, which will do whatever custom emission the specific `OpOverride` implements. If there is no entry in the map, codegen continues through the normal path, and none of the `Emit*()` functions need to have special handling for format strings.

This change also reorganizes codegen, removing places where all nodes are re-iterated to emit special operations, like assert or trace. This seems to improve codegen performance significantly.

PiperOrigin-RevId: 462043641