Adding labels to DSLX asserts.
It's common for asserts in SystemVerilog to be associated with a name/label, so DSLX needs to support their specification in fail! "macros". This change adds support for that, but does not yet connect them to the IR. That'll come next: I didn't want this CL to get too big.

PiperOrigin-RevId: 463174121