Updates top-level proc "boundary" channels to use their param names instead of generated ones.
The channels used as the interface to top-level procs (i.e., the "boundary" channels) previously used auto-generated names, which are rather obscure (based on instantiation order and invocation span). This made it hard to refer to those channels on the other side of the interface.

This change makes it so that the boundary channels have the same name as their declarations. For example, for the following proc,

```
proc foo {
  config(a: chan in u32, b: chan out u32) { ... }
  next(tok: token) { ... }
}
```

There would be two boundary channels declared, named "a", and "b".

Resolves #656.

PiperOrigin-RevId: 467030988