Enabling specification of initial state for top-level procs.
This change enables DSLX procs that maintain state to be converted to IR. This requires specification of a new flag, "--top_proc_initial_state", that contains the string representation of the IR values to assign to the Proc's initial state.

This change also clarifies the structure of a Proc's carried state: previously, there was a mixture of treating it as a series of values or treating it as a singular tuple of values. This change rigorously defines it as the latter There's no reason it can't be changed to the former in the future: this simply seemed better while working on the change.

Also contains miscellaneous cleanups encountered along the way, particularly removing the parsing wrinkle that required the recurrent state to be wrapped in an unnecessary tuple, e.g., a single tuple state value no longer needs to be wrapped in another tuple at next function termination.

PiperOrigin-RevId: 467270274